refactor: rename oauth device auth files
